refactor: Operator Enum 추가

+++ src/main/kotlin/calculator/ArithmeticExpression.kt
@@ -7,7 +7,7 @@ data class ArithmeticExpression(
 ) {
     val operators: List<String>
 
-    val operands: List<Int>
+    val operands: List<Double>
 
     init {
         require(expression.isNotBlank() && expression.isNotEmpty()) {"연산식은 공백이 될수 없습니다."}
@@ -15,7 +15,7 @@ data class ArithmeticExpression(
         val split = expression.split(DELIMITER)
         operators = split.filterIndexed { i, _ ->  i % 2 == 1}
         operands = split.filterIndexed { i, _ ->  i % 2 == 0}
-                        .map { s: String ->  s.toInt() }
+                        .map { s: String ->  s.toDouble() }
     }
 
 }
